now making the servives of  the project

+++ src/main/java/com/nitish/insta/Service/UsersService.java
@@ -7,11 +7,11 @@ import java.util.List;
 
 public interface UsersService {
     UsersDto registerNewUser(UsersDto user,Device device);
-    UsersDto updateUser(UsersDto user,Integer userId);
-    UsersDto getUser(Integer userId);
+    UsersDto updateUser(UsersDto user,Long userId);
+    UsersDto getUser(Long userId);
     void loginUser(Device device,String email);
     List<UsersDto> getAllUsers();
-    void deleteUser(Integer userId);
+    void deleteUser(Long userId);
     void googleLoginRegistration(String email,String providerId,String username,Device device);
     String changeUserName(String userName,String email);
     String changeFullName(String fullName,String email);
